Added admin verification, albeit its just the implimentation.

+++ frontend/src/Components/ProductCard.jsx
@@ -13,32 +13,54 @@ export default function ProductCard({
   handleEditProduct,
   _id,
   handleDeleteProduct,
+  isAdmin,
 }) {
-  return (
-    <div className="ProductCard">
-      <h3>{productName}</h3>
-      <img src={image} alt="" />
-      <h4>{brand}</h4>
-      <QuantityCounter
-        handleAddQuantity={handleAddQuantity}
-        productQuantity={productQuantity}
-        handleRemoveQuantity={handleRemoveQuantity}
-        id={id}
-        mode="product"
-      />
-      <h3>{price}</h3>
-      <button onClick={() => handleAddToCart(id)}>Add to Cart</button>
-      <button
-        id="edit-button"
-        onClick={() =>
-          handleEditProduct({ price, brand, productName, image, _id })
-        }
-      >
-        Edit
-      </button>
-      <button className="RemoveButton" onClick={() => handleDeleteProduct(_id)}>
-        Delete
-      </button>
-    </div>
-  );
+  if(isAdmin){
+    return (
+      <div className="ProductCard">
+        <h3>{productName}</h3>
+        <img src={image} alt="" />
+        <h4>{brand}</h4>
+        <QuantityCounter
+          handleAddQuantity={handleAddQuantity}
+          productQuantity={productQuantity}
+          handleRemoveQuantity={handleRemoveQuantity}
+          id={id}
+          mode="product"
+        />
+        <h3>{price}</h3>
+        <button onClick={() => handleAddToCart(id)}>Add to Cart</button>
+        <button
+          id="edit-button"
+          onClick={() =>
+            handleEditProduct({ price, brand, productName, image, _id })
+          }
+        >
+          Edit
+        </button>
+        <button className="RemoveButton" onClick={() => handleDeleteProduct(_id)}>
+          Delete
+        </button>
+      </div> 
+    );
+  }
+  else{
+    return (
+      <div className="ProductCard">
+        <h3>{productName}</h3>
+        <img src={image} alt="" />
+        <h4>{brand}</h4>
+        <QuantityCounter
+          handleAddQuantity={handleAddQuantity}
+          productQuantity={productQuantity}
+          handleRemoveQuantity={handleRemoveQuantity}
+          id={id}
+          mode="product"
+        />
+        <h3>{price}</h3>
+        <button onClick={() => handleAddToCart(id)}>Add to Cart</button>
+      </div>
+    )
+  }
+  
 }
